initial checkin.
pulled from another project into its own.  haven't cleaned it up much yet, but it works well in the general projects it was in.
first file, this one, deals with creating multi-threaded apps similar to CUDA but using multiple core.  tested with a raytracer project, which required only two lines of code change to use this class.