wrote and tested modifications to support new workloads
This change (and test for it) allows for workload to be generated on the
main thread.  The threads underneath will start doing work ASAP, and will
be notified when all of the workloads have been added to the list.

Usage case:  if the main thread is decoding a stream, as it finishes a
packet, a thread can then handle that packet while the next packet is
being decoded on the main thread.  This allows for the masking of decode
time - no waiting until everything is decoded before firing off threads.

Base implementation of it this new API.  Need to clean it up a bit before
merging it into main (and verifying it on Mac, etc).