added missing locks
predicate variables for the wait function were being accessed outside
of a lock, which means there was a slight chance of it being modified
after being checked by the wait's predicate check, but before the
actual wait() was called.

added lock guards around all access to the predicate values to prevent
this possible endless wait for a notify that had already happened.